+++ src/test/examples/this.ts
@@ -1,0 +1,5 @@
+function doThings(this: void, str: string) {
+  return str;
+}
+
+doThings("what?");